+++ src/slurmd/job.c
@@ -105,7 +105,7 @@ job_create(launch_tasks_request_msg_t *msg, slurm_addr *cli_addr)
 	debug3("entering job_create");
 
 	if ((pwd = _pwd_create((uid_t)msg->uid)) < 0) {
-		error("uid %ld not found on system", msg->uid);
+		error("uid %ld not found on system", (long) msg->uid);
 		return NULL;
 	}
 	job = xmalloc(sizeof(*job));
@@ -188,7 +188,7 @@ job_batch_job_create(batch_job_launch_msg_t *msg)
 	uint32_t      global_taskid = 0;
 
 	if ((pwd = _pwd_create((uid_t)msg->uid)) < 0) {
-		error("uid %ld not found on system", msg->uid);
+		error("uid %ld not found on system", (long) msg->uid);
 		return NULL;
 	}
 
@@ -459,5 +459,5 @@ job_delete_shm(slurmd_job_t *job)
 {
 	if (shm_delete_step(job->jobid, job->stepid) == SLURM_FAILURE)
 		error("deleting step:  %ld.%ld not found in shmem", 
-				job->jobid, job->stepid); 
+				(long) job->jobid, (long) job->stepid); 
 }
